Added files and reorganized directories
Added buell extra credit
Added links to git-info file
Added links from dropbox.cse.sc.edu:
	* mobaxterm zip file
	* ssh client exe
	* remote login instructions
	* text file with further links

Sorted files in main directory:
	* Files hosted on https://cse.sc.edu/~pokeefe/unix-linux/resources/
	  have been relocated to from_website
	* Added directory for pdf files

Site https://cse.sc.edu/~pokeefe/unix-linux/resources/
no longer is necessary (as all of its resources are now hosted in
from_website
