+++ c++/extra_credit_buell/instructions.okeefe.txt
@@ -1,0 +1,19 @@
+CSCE 215 Fall 2017 final Extra Credit
+
+-------------------------------------
+Suggested steps to follow:
+1. uncompress and untar the gz file
+2. locate the README and read it.
+3. mkdir mydirectory/lastfirst_hwEC
+4. cp Main.cpp to your lastfirst_hwEC dir
+5. run scripts
+6. check the ouput in testdirectory/lastfirst_hwEC
+7. modify Main.cpp
+8. re-run scripts
+9. if testing complete submit the .gz file in assignment dir
+
+--------------------
+EC 1pt
+do the above correctly. modify source to print out your name.
+--------------------
+
