+++ from_website/numCmdArgs.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+echo "first cmd line variable is:"
+echo $1
+echo "second cmd line varibal is:"
+echo $2
+echo "number of cmd line variables: "
+echo $#
+echo "command is: "
+echo $0
