+++ git-info
@@ -10,3 +10,11 @@ Git and bitbucket tutorial:
 https://bitbucket.org/okeefe/tutorial
 
 
+Reference manual:
+
+https://git-scm.com/book/en/v2
+
+
+Github tutorials:
+
+https://help.github.com/
